+++ CHANGELOG.md
@@ -28,6 +28,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 - Update `nginx` to 1.26.2.
 - Deps: update `@angular` to 18.2.1.
 - Deps: update `marked` to 14.1.0.
+- Deps (dev): update `@types/node` to 20.16.2.
 
 ### Fixed